add signout button and session states to navbar, abstract the component tree, ensure navbar is a server component
